Cambio de int a double para más precicón en resultado y calcula el area del circulo

+++ src/main/java/my/com/vmbl/poo/figuras/Figura.java
@@ -9,5 +9,5 @@ package my.com.vmbl.poo.figuras;
  * @author victor-briseno
  */
 public interface Figura {
-    int calcularArea();
+    double calcularArea();
 }
